requirements.txt
openai>=1.18.0
openai-agents>=0.0.14
pinecone>=2.3.1
streamlit>=1.34.0
python-dotenv
